refactor(cli): inline device-flow into auth command, add saveFromLogin to store
Delete device-flow.ts — it was single-use glue between the API client
and session store. The mapping from API response types to AuthSession
now lives in store.saveFromLogin(), and the polling loop lives in the
login command handler where it belongs.
